Update to the main map
Fixes all of the map issues listed on github. Removed the useless hallway between the science server room, and the sec checkpoint making the server room a little larger. Fixed the naming issue with the sec points windoors. Added plating beneath the RD's windows. Fixed the medical supply room vent, and viro vent. Added a air alarm to the medical supply room. Fixed the naming issue with the door. Locked the viro's APC.  Added a new vent to the hallway outside tool storage. Moved the scrubbers over one in the dorms. Added building permits to the engineers lockers.